fix: replace deprecated (double) cast with (float)
PHP 8.4 deprecates the non-canonical (double) cast. Replace all 11
instances with the canonical (float) equivalent. No behavioral change.

+++ lib/functions.php
@@ -9545,7 +9545,7 @@ if (!function_exists('stats_standard_deviation')) {
 		$carry = 0.0;
 
 		foreach ($items as $val) {
-			$d = ((double) $val) - $mean;
+			$d = ((float) $val) - $mean;
 			$carry += $d * $d;
 		}
